+++ app/celery/scheduled_tasks.py
@@ -5,15 +5,12 @@ from datetime import (
 
 from celery.signals import worker_process_shutdown
 from flask import current_app
-from notifications_utils.recipients import RecipientCSV
-from sqlalchemy import or_, and_
+from sqlalchemy import and_
 from sqlalchemy.exc import SQLAlchemyError
 from notifications_utils.s3 import s3upload
 
 from app.aws import s3
 from app import notify_celery
-from app.celery import celery
-from app.dao.templates_dao import dao_get_template_by_id
 from app.performance_platform import total_sent_notifications, processing_time
 from app import performance_platform_client
 from app.dao.date_util import get_month_start_and_end_date_in_utc
@@ -22,8 +19,8 @@ from app.dao.invited_user_dao import delete_invitations_created_more_than_two_da
 from app.dao.jobs_dao import (
     dao_get_letter_job_ids_by_status,
     dao_set_scheduled_jobs_to_pending,
-    dao_get_jobs_older_than_limited_by,
-    dao_get_job_by_id)
+    dao_get_jobs_older_than_limited_by
+)
 from app.dao.monthly_billing_dao import (
     get_service_ids_that_need_billing_populated,
     create_or_update_monthly_billing
@@ -44,7 +41,6 @@ from app.dao.provider_details_dao import (
 from app.dao.users_dao import delete_codes_older_created_more_than_a_day_ago
 from app.models import (
     Job,
-    Notification,
     LETTER_TYPE,
     JOB_STATUS_READY_TO_SEND,
     JOB_STATUS_IN_PROGRESS
@@ -53,10 +49,7 @@ from app.notifications.process_notifications import send_notification_to_queue
 from app.statsd_decorators import statsd
 from app.celery.tasks import (
     create_dvla_file_contents_for_notifications,
-    get_template_class,
-    process_job,
-    process_row,
-    job_complete
+    process_job
 )
 from app.config import QueueNames, TaskNames
 from app.utils import convert_utc_to_bst
@@ -409,44 +402,3 @@ def check_job_status():
             queue=QueueNames.JOBS
         )
         raise JobIncompleteError("Job(s) {} have not completed.".format(job_ids))
-
-
-@notify_celery.task(name='process-incomplete-jobs')
-@statsd(namespace="tasks")
-def process_incomplete_jobs(job_ids):
-    current_app.logger.info("Resuming Job(s) {}".format(job_ids))
-    for job_id in job_ids:
-        process_incomplete_job(job_id)
-
-
-def process_incomplete_job(job_id):
-
-    job = Job.query.filter(Job.id == job_id).one()
-
-    last_notification_added = Notification.query.filter(
-        Notification.job_id == job_id
-    ).order_by(
-        Notification.job_row_number.desc()
-    ).first()
-
-    if last_notification_added:
-        resume_from_row = last_notification_added.job_row_number
-    else:
-        resume_from_row = -1  # The first row in the csv with a number is row 0
-
-    current_app.logger.info("Resuming job {} from row {}".format(job_id, resume_from_row))
-
-    db_template = dao_get_template_by_id(job.template_id, job.template_version)
-
-    TemplateClass = get_template_class(db_template.template_type)
-    template = TemplateClass(db_template.__dict__)
-
-    for row_number, recipient, personalisation in RecipientCSV(
-            s3.get_job_from_s3(str(job.service_id), str(job.id)),
-            template_type=template.template_type,
-            placeholders=template.placeholders
-    ).enumerated_recipients_and_personalisation:
-        if row_number > resume_from_row:
-            process_row(row_number, recipient, personalisation, template, job, job.service)
-
-    job_complete(job, job.service, template, True)
